# Refactor styling system to use conventional Tailwind theming approach