--- a/components/ui/custom-select.tsx
+++ b/components/ui/custom-select.tsx
@@ -43,7 +43,7 @@ export default function CustomSelect({
         value={value}
         onChange={handleChange}
         disabled={disabled}
-        className="w-full min-h-[44px] bg-background border border-border rounded-md px-4 py-2 text-left font-mono text-xs focus:outline-none focus:ring-2 focus:ring-yellow-500 focus:border-yellow-500 disabled:opacity-50 disabled:cursor-not-allowed hover:border-muted-foreground transition-colors touch-manipulation appearance-none cursor-pointer text-foreground pr-10"
+        className="w-full min-h-[44px] bg-background border border-border rounded-md px-4 py-2 text-left font-mono text-xs focus:outline-none focus:ring-2 focus:ring-focus-ring focus:border-focus-ring disabled:opacity-50 disabled:cursor-not-allowed hover:border-muted-foreground transition-colors touch-manipulation appearance-none cursor-pointer text-foreground pr-10"
       >
         {placeholder && (
           <option value="" disabled hidden>