… components/mobile/progressive-disclosure-panel.tsx
… components/ui/collapsible-control-group.tsx
… components/ui/compact-color-picker.tsx
… components/ui/grouped-simulation-controls-panel.tsx
… components/ui/simulation-controls-panel.tsx
… components/ui/theme-toggle.test.tsx
… components/ui/theme-toggle.tsx
… components/ui/viewport-constrained-panel.tsx